requirements.txt
Jinja2~=2.11.2
requests~=2.24.0
beautifulsoup4~=4.9.1
python-wordpress-xmlrpc~=2.3
qiniu
fire~=0.3.1
wheel~=0.34.2